로그포맷에 맞는 필터링 가능하게 구현

+++ web/src/main/java/web/web/admin/repository/LogFormatRepository.java
@@ -3,5 +3,5 @@ package web.web.admin.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.web.admin.domain.LogFormat;
 
-public interface LogFormatRepository extends JpaRepository<LogFormat, Integer> {
+public interface LogFormatRepository extends JpaRepository<LogFormat, Long> {
 }
